refactor(core): split operation element parsing responsibilities
Why:
- 中凝集だった operation reader に配列オーケストレーションと要素パース責務が同居していたため。

What Changed:
- Validate/Execute の operation reader から 1 要素の検証・読取処理を専用の element reader へ分離。
- operation reader は配列走査と重複 ID 検知のみを担うように簡素化。
- 同期処理の operation reader から不要な CancellationToken 連携を削除し、呼び出し側を更新。
- RequestSchemaPolicy を受ける引数に in 修飾を付与し、意図を明確化。

Impact:
- 既存の契約バリデーション挙動を維持したまま、可読性と保守性を改善。
